+++ modules/se_sdk3_hosting/Controller.cpp
@@ -2185,7 +2185,7 @@ void MpController::ImportBankXml(const char* xmlfilename)
 				formatVersion = 1;
 				try
 				{
-					fourCC = std::stoul(hexcode, nullptr, 16);
+					fourCC = static_cast<int32_t>(std::stoul(hexcode, nullptr, 16));
 				}
 				catch (...)
 				{
